fail

+++ Hello.java
@@ -1,7 +1,7 @@
 public class Hello {
     public static void main(String[] args) {
-        int a = 5;
-        double c = 11;
+        double a = 5;
+        int c = 11;
         c = a+c;
         System.out.println(a);
         System.out.println("Hello, Java!!!!!!! AND GIT !!!");
